+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 * Fix validation error with END DATE for `Cash drawer reconciliation report` modal. Refs UIU-2175.
 * Handle empty report data in `CashDrawerReconciliationReportCSV`. Refs UIU-2184.
 * Fee/fine owners not available for selection on Transfer Fee/Fine modal. Refs UIU-2174.
+* Delete user with check for open transactions. Refs UIU-1971.
 * Save the patron expiration with the time set to 23:59:59. Refs UIU-2182.
 
 ## [6.0.6](https://github.com/folio-org/ui-users/tree/v6.0.6) (2021-06-17)